doc update and pointer moving
I moved some of the object pointer storage code from random methods back into the class responsbile for what is being stored.

I also added the ability to store a reference time in the plot which this class can retrieve. This is useful for creating multiple plots which are aligned based on some global start time.